+++ test/unit/maze/option/processor_test.rb
@@ -37,8 +37,6 @@ class ProcessorTest < Test::Unit::TestCase
     assert_equal '1.2.3.4', config.bind_address
     assert_equal 1234, config.port
     assert_equal 'root', config.document_server_root
-    assert_equal '5.6.7.8', config.document_server_bind_address
-    assert_equal 5678, config.document_server_port
     assert_false config.start_appium
     assert_equal '123456789012345678901234567890ab', config.bugsnag_repeater_api_key
     assert_equal '000006789012345678901234567890ab', config.hub_repeater_api_key
@@ -58,8 +56,6 @@ class ProcessorTest < Test::Unit::TestCase
     assert_equal 'ABC', config.apple_team_id
     assert_equal '123', config.device_id
     assert_equal 'root', config.document_server_root
-    assert_equal '5.6.7.8', config.document_server_bind_address
-    assert_equal 5678, config.document_server_port
     assert_false config.start_appium
 
     assert_nil config.os_version
@@ -87,8 +83,6 @@ class ProcessorTest < Test::Unit::TestCase
     assert_equal nil, config.bind_address
     assert_equal 9339, config.port
     assert_equal nil, config.document_server_root
-    assert_equal nil, config.document_server_bind_address
-    assert_equal 9340, config.document_server_port
 
     assert_equal :id, config.locator
     assert_nil config.capabilities
@@ -115,8 +109,6 @@ class ProcessorTest < Test::Unit::TestCase
     assert_equal 1234, config.port
 
     assert_equal nil, config.document_server_root
-    assert_equal nil, config.document_server_bind_address
-    assert_equal 9340, config.document_server_port
 
     assert_equal :id, config.locator
     assert_nil config.capabilities
@@ -180,15 +172,4 @@ class ProcessorTest < Test::Unit::TestCase
     ENV.delete('MAZE_HUB_REPEATER_API_KEY')
   end
 
-  def test_environment_options
-    ENV['USE_LEGACY_DRIVER'] = '1'
-    ENV['MAZE_PORT'] = '1234'
-
-    args = []
-    options = Maze::Option::Parser.parse args
-    config = Maze::Configuration.new
-    Maze::Option::Processor.populate config, options
-
-    assert_true config.legacy_driver?
-  end
 end